requirements.txt
torch
sentence-transformers
chromadb
requests
python-docx
fastapi
pydantic
uvicorn
sqlalchemy>=2.0
aiosqlite
passlib[bcrypt]
argon2-cffi
pydantic-settings
email-validator
python-jose[cryptography]